requirements.txt
google-genai>=0.1.0
Pillow>=9.0.0
python-dotenv>=1.0.0
